1.67

+++ index.js
@@ -1,7 +1,14 @@
 const express = require("express")
 const app = express()
 const port = 24555
-const { Messages, Users, Sessions, Friends, Feedback } = require("./models")
+const {
+  Messages,
+  Users,
+  Sessions,
+  Friends,
+  Feedback,
+  Chats
+} = require("./models")
 const rateLimit = require("express-rate-limit")
 const argon2 = require("argon2")
 const cryptoRandomString = require("crypto-random-string")
@@ -83,8 +90,14 @@ app.use(
   })
 )
 
-app.get("/api/messages", auth, async (req, res) => {
+app.get("/api/messages/:chatId", auth, async (req, res) => {
+  if (!req.params.chatId) {
+    return res.status(400).json({
+      message: "Missing chatId"
+    })
+  }
   const messages = await Messages.findAll({
+    where: { chatId: req.params.chatId },
     include: [
       {
         model: Users,
@@ -93,7 +106,33 @@ app.get("/api/messages", auth, async (req, res) => {
       }
     ]
   })
-  res.json(messages)
+  return res.json(messages)
+})
+
+app.get("/api/chat/:chatId", auth, async (req, res) => {
+  const chat = await Chats.findOne({
+    where: {
+      id: req.params.chatId
+    }
+  })
+  if (!chat) {
+    res.status(400)
+    res.json({
+      message: "Chat does not exist"
+    })
+    return
+  }
+  chat.dataValues.messages = await Messages.findAll({
+    where: { chatId: req.params.chatId },
+    include: [
+      {
+        model: Users,
+        as: "user",
+        attributes: ["id", "username", "avatar"]
+      }
+    ]
+  })
+  return res.json(chat)
 })
 
 app.get("/api/users", auth, async (req, res) => {
@@ -108,6 +147,13 @@ app.get("/api/users", auth, async (req, res) => {
     attributes: ["id", "username", "avatar", "status", "statusMessage"]
   })
   res.json(users)
+})
+
+app.get("/api/chats", auth, async (req, res) => {
+  const chats = await Chats.findAll({
+    attributes: ["id", "name", "description", "icon", "owner", "latest"]
+  })
+  res.json(chats)
 })
 
 app.get("/api/user", auth, async (req, res) => {
@@ -206,28 +252,62 @@ app.post("/api/message", auth, async (req, res) => {
       })
       return
     }
+    if (!req.body.chatId) {
+      res.status(400)
+      res.json({
+        message: "Chat not specified"
+      })
+      return
+    }
+    const chat = await Chats.findOne({
+      where: {
+        id: req.body.chatId
+      }
+    })
+    if (!chat) {
+      res.status(400)
+      res.json({
+        message: "Chat does not exist"
+      })
+      return
+    }
+    if (chat.requireVerification && !req.user.emailVerified) {
+      res.status(400)
+      res.json({
+        message: "User not verified"
+      })
+      return
+    }
     const messageText = req.body.messageContents.trim()
     const replyMessage = req.body?.reply
     if (messageText) {
       const message = await Messages.create({
         messageContents: messageText,
         userName: req.user.id,
-        reply: replyMessage
+        reply: replyMessage,
+        chatId: req.body.chatId
       })
-      resolveEmbeds(req, message)
-        .then(async () => {
-          const messages = await Messages.findAll({
-            include: [
-              {
-                model: Users,
-                as: "user",
-                attributes: ["id", "username", "avatar"]
-              }
-            ]
-          })
-          res.json(messages)
-        })
-        .catch(async () => {})
+      resolveEmbeds(req, message).catch(async () => {})
+      await chat.update({
+        latest: Date.now()
+      })
+      console.log("test")
+      chat.dataValues.messages = await Messages.findAll({
+        where: { chatId: req.body.chatId },
+        include: [
+          {
+            model: Users,
+            as: "user",
+            attributes: ["id", "username", "avatar"]
+          }
+        ]
+      })
+      console.log("test")
+      const chats = await Chats.findAll({
+        attributes: ["id", "name", "description", "icon", "owner", "latest"]
+      })
+      const data = { chat, chats }
+      res.json(data)
     }
   } catch (e) {
     res.status(500)
@@ -235,6 +315,57 @@ app.post("/api/message", auth, async (req, res) => {
       message: "Something went wrong" + e
     })
   }
+})
+
+app.post("/api/create-chat", auth, async (req, res) => {
+  if (!req.body.name) {
+    res.status(400)
+    res.json({
+      message: "Chat name not specified"
+    })
+    return
+  }
+  if (typeof req.body.requireVerification !== "boolean") {
+    res.status(400)
+    res.json({
+      message: "requireVerification not specified"
+    })
+    return
+  }
+  if (req.body.name.length > 30) {
+    res.status(400)
+    res.json({
+      message: "Chat name too long"
+    })
+    return
+  }
+  if (req.body.description.length > 500) {
+    res.status(400)
+    res.json({
+      message: "Chat description too long"
+    })
+    return
+  }
+  if (!req.user.admin) {
+    res.status(403)
+    res.json({
+      message: "Forbidden"
+    })
+    return
+  }
+  const chat = await Chats.create({
+    name: req.body.name,
+    description: req.body.description,
+    icon: req.body.icon,
+    owner: req.user.id,
+    requireVerification: req.body.requireVerification,
+    latest: Date.now()
+  })
+  const chats = await Chats.findAll({
+    attributes: ["id", "name", "description", "icon", "owner", "latest"]
+  })
+  let data = { chat, chats }
+  res.json(data)
 })
 
 app.post("/api/register", async (req, res) => {
@@ -563,6 +694,7 @@ app.post("/api/resend-verification", auth, async (req, res) => {
     .catch((error) => {
       console.log("Error occurred while sending email:", error)
     })
+  return res.sendStatus(204)
 })
 
 app.post("/api/verify", auth, async (req, res) => {
@@ -594,12 +726,64 @@ app.post("/api/verify", auth, async (req, res) => {
 })
 
 app.delete("/api/delete/:messageId", auth, async (req, res) => {
+  const message = await Messages.findOne({
+    where: {
+      id: req.params.messageId
+    }
+  })
+  if (!message) {
+    res.status(400)
+    res.json({
+      message: "Message does not exist"
+    })
+    return
+  }
   const where = req.user.admin
     ? { id: req.params.messageId }
     : { id: req.params.messageId, userName: req.user.id }
-
   await Messages.destroy({ where })
-  res.sendStatus(204)
+  const messages = await Messages.findAll({
+    where: { chatId: message.chatId },
+    include: [
+      {
+        model: Users,
+        as: "user",
+        attributes: ["id", "username", "avatar"]
+      }
+    ]
+  })
+  res.json(messages)
+})
+
+app.delete("/api/delete-chat/:chatId", auth, async (req, res) => {
+  const chat = await Chats.findOne({
+    where: {
+      id: req.params.chatId
+    }
+  })
+  if (!chat) {
+    res.status(400)
+    res.json({
+      message: "Chat does not exist"
+    })
+    return
+  }
+  if (!req.user.admin || chat.owner !== req.user.id) {
+    res.status(403)
+    res.json({
+      message: "Forbidden"
+    })
+    return
+  }
+  await Chats.destroy({
+    where: {
+      id: req.params.chatId
+    }
+  })
+  const chats = await Chats.findAll({
+    attributes: ["id", "name", "description", "icon", "owner", "latest"]
+  })
+  res.json(chats)
 })
 
 app.delete("/api/delete-feedback/:feedbackId", auth, async (req, res) => {
